Moved from sensor to platform
